#12935: Refactor sweep_framework structure and naming
