+++ .github/workflows/all-static-checks.yaml
@@ -57,4 +57,4 @@ jobs:
       - name: Check sweeps workflow option count against sweep file count
         run: |
           pip install pyyaml
-          python tests/sweep_framework/sweeps_workflow_verification.py
+          python tests/sweep_framework/framework/sweeps_workflow_verification.py
